requirements.txt
jupyterlab>=4.4.3
jupyterlab-git>=0.44.0
jupyter-app-launcher>=0.1.0
dotenv~=0.9.9
voila~=0.5.8
OpenAI~=1.97.0
langgraph~=0.5.3
langchain-nvidia-ai-endpoints~=0.3.12
langchain-openai~=0.2.0
pydantic~=2.11.7
tavily-python~=0.7.10
flask>=3.0.0

